move archive facility to facility settings in portfolio
